backend/storage-access-function/requirements.txt
functions-framework==3.*
google-cloud-storage==3.2.0
Flask==3.0.*
